[core] [no-op] Cleanup utils folder bazel dependency (#49938)
