+++ src/ray/common/asio/asio_util.h
@@ -23,6 +23,7 @@
 
 #include "ray/common/asio/instrumented_io_context.h"
 #include "ray/util/array.h"
+#include "ray/util/thread_utils.h"
 #include "ray/util/util.h"
 
 template <typename Duration>
