+++ src/ray/rpc/grpc_server.cc
@@ -23,6 +23,7 @@
 #include "ray/common/ray_config.h"
 #include "ray/rpc/common.h"
 #include "ray/stats/metric.h"
+#include "ray/util/thread_utils.h"
 #include "ray/util/util.h"
 
 namespace ray {
