+++ src/ray/util/event.cc
@@ -21,6 +21,9 @@
 
 #include "absl/base/call_once.h"
 #include "absl/time/time.h"
+#include "ray/util/random.h"
+#include "ray/util/string_utils.h"
+#include "ray/util/timestamp_utils.h"
 
 using json = nlohmann::json;
