+++ src/ray/util/logging.cc
@@ -42,7 +42,7 @@
 #include "nlohmann/json.hpp"
 #include "ray/util/event_label.h"
 #include "ray/util/filesystem.h"
-#include "ray/util/util.h"
+#include "ray/util/thread_utils.h"
 #include "spdlog/sinks/basic_file_sink.h"
 #include "spdlog/sinks/rotating_file_sink.h"
 #include "spdlog/sinks/stdout_color_sinks.h"
