+++ src/ray/util/random.h
@@ -1,4 +1,4 @@
-// Copyright 2017 The Ray Authors.
+// Copyright 2025 The Ray Authors.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,16 +14,17 @@
 
 #pragma once
 
-#include <cstdint>
+#include "absl/random/random.h"
 
-namespace ray {
+/// A helper function to fill random bytes into the `data`.
+/// Warning: this is not fork-safe, we need to re-seed after that.
+template <typename T>
+void FillRandom(T *data) {
+  RAY_CHECK(data != nullptr);
 
-// A helper function for doing memcpy with multiple threads. This is required
-// to saturate the memory bandwidth of modern cpus.
-void parallel_memcopy(uint8_t *dst,
-                      const uint8_t *src,
-                      int64_t nbytes,
-                      uintptr_t block_size,
-                      int num_threads);
-
-}  // namespace ray
+  thread_local absl::BitGen generator;
+  for (size_t i = 0; i < data->size(); i++) {
+    (*data)[i] = static_cast<uint8_t>(
+        absl::Uniform(generator, 0, std::numeric_limits<uint8_t>::max()));
+  }
+}
