+++ src/ray/util/timestamp_utils.h
@@ -1,4 +1,4 @@
-// Copyright 2017 The Ray Authors.
+// Copyright 2025 The Ray Authors.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,16 +14,14 @@
 
 #pragma once
 
-#include <cstdint>
+#include <chrono>
 
 namespace ray {
 
-// A helper function for doing memcpy with multiple threads. This is required
-// to saturate the memory bandwidth of modern cpus.
-void parallel_memcopy(uint8_t *dst,
-                      const uint8_t *src,
-                      int64_t nbytes,
-                      uintptr_t block_size,
-                      int num_threads);
+inline int64_t current_sys_time_s() {
+  std::chrono::seconds s_since_epoch = std::chrono::duration_cast<std::chrono::seconds>(
+      std::chrono::system_clock::now().time_since_epoch());
+  return s_since_epoch.count();
+}
 
 }  // namespace ray
